% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catalysis}
\alias{catalysis}
\title{Single-Atom Catalysis Data}
\format{
A list with 4 objects:
\describe{
\item{X}{Primary feature matrix: physical properties of transition metals and oxide supports}
\item{y}{Reponse variable: binding energy of metal/oxide pairs}
\item{head}{Column names of X}
\item{unit}{Unit of columns of X}
}
}
\usage{
catalysis
}
\description{
Single-Atom Catalysis Data
}
\keyword{datasets}
