% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iBART_real_data}
\alias{iBART_real_data}
\title{iBART Real Data Result}
\format{
A list of iBART outputs
\describe{
\item{iBART_model}{A cv.glmnet object storing the iBART selected model}
...
}
}
\usage{
iBART_real_data
}
\description{
iBART result in the real data vignette
}
\keyword{datasets}
