% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iBART_sim}
\alias{iBART_sim}
\title{iBART Simulation Result}
\format{
A list of iBART outputs
\describe{
\item{iBART_model}{A cv.glmnet object storing the iBART selected model}
...
}
}
\usage{
iBART_sim
}
\description{
iBART result in the simulation vignette
}
\keyword{datasets}
